### 2019-05-17 #1455 @arfon
Hi @medewitt, thanks for your submission to JOSS. Could you please update the title field of you paper to be (note the double quotations wrapping the string):

```
"tidyqwi: A Tidy Approach to Accessing The US Census Bureau's Quarterly Workforce Indicators"
```
