### 2019-05-17 69fe7c5
Update paper.md
reformatted paper title per submission guidelines
