### 2019-05-17 #1455 @medewitt
@tweed1e , @koulvi , @zedtaha might be potentially good reviewers for this package (economics/ econometrics in subject matter).
