### 2019-05-20 #1455 @danielskatz
> whedon typo

btw, you can assign yourself as an editor too :)
