### 2019-05-20 #1455 @alexhanna
> btw, you can assign yourself as an editor too :)

I did not know this! Thanks. :)
