### 2019-05-20 #1462 @whedon
[REVIEW]: tidyqwi: A Tidy Approach to Accessing The US Census Bureau's Quarterly Workforce Indicators
**Submitting author:** @medewitt (<a href="http://orcid.org/0000-0001-8940-1967">Michael DeWitt</a>)
**Repository:** <a href="https://github.com/medewitt/tidyqwi" target ="_blank">https://github.com/medewitt/tidyqwi</a>
**Version:** v0.1.0
