### 2019-05-20 #1455 @alexhanna
Noted!

On Mon, May 20, 2019 at 8:57 AM Michael DeWitt <notifications@github.com>
wrote:

> @alexhanna <https://github.com/alexhanna> , I just saw @koulvi on the
> googlesheet list of possible reviewers with expertise in econ and R. Must
> not be a valid user anymore.
>
> Also, thanks for editing.
>
