### 2019-05-20 #1455 @arfon
> It would be good to find a second reviewer as well, unless you are very well-versed in the area of this software (which might be the case...) We strongly prefer to have 2 reviewers.

@danielskatz ping me when you've added the second review and I'll edit the review checklist and invite the second reviewer to the review repo (so they can update their checklist).
