### 2019-05-31 medewitt/tidyqwi#21 @tweed1e
Possible to consolidate license files?
RE: JOSS review https://github.com/openjournals/joss-reviews/issues/1462, is there a way to combine the LICENSE and LICENSE.md files into one plain-text (.md or other) license file?

I'm not sure what the CRAN license file requirements are, but LICENSE doesn't say what type of license it is, and LICENSE.md says MIT, so one consistent file would avoid the confusion 🙏.
