### 2019-06-03 medewitt/tidyqwi#21 @medewitt
@tweed1e per CRAN [guidelines ](https://cran.r-project.org/web/packages/policies.html)it looks like I can remove the LICENSE.md file as long as I have the license text in the DESCRIPTION file [(License: MIT + License)](https://www.r-project.org/Licenses/MIT) and have the LICENSE (plain text with copyright year)(more for printing the license text on github). I'll go ahead and remove the markdown license version.
