### 2019-06-03 18c8ec7
removed license.md file in order to make license situation more clear. PAsses all CRAN checks
