### 2019-06-06 8f364e7
updated website again andf install methods
