+++ README.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 <!-- badges here -->
 
 [![CRAN status](https://www.r-pkg.org/badges/version/tidyqwi)](https://cran.r-project.org/package=tidyqwi)
-[![Travis build status](https://travis-ci.org/medewitt/tidyqwi.svg?branch=master)](https://travis-ci.org/medewitt/tidyqwi) [![Coverage status](https://codecov.io/gh/medewitt/tidyqwi/branch/master/graph/badge.svg)](https://codecov.io/github/medewitt/tidyqwi?branch=master) [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental) [![](https://cranlogs.r-pkg.org/badges/tidyqwi)](https://cran.r-project.org/package=tidyqwi)
+[![Travis build status](https://travis-ci.org/medewitt/tidyqwi.svg?branch=master)](https://travis-ci.org/medewitt/tidyqwi) [![Coverage status](https://codecov.io/gh/medewitt/tidyqwi/branch/master/graph/badge.svg)](https://codecov.io/github/medewitt/tidyqwi?branch=master) [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental) [![](https://cranlogs.r-pkg.org/badges/tidyqwi)](https://cran.r-project.org/package=tidyqwi)[![status](http://joss.theoj.org/papers/96bac49d6912cea4d191abff922d7c2e/status.svg)](http://joss.theoj.org/papers/96bac49d6912cea4d191abff922d7c2e)
 
 <!-- end badges -->
 
@@ -28,9 +28,19 @@ This is a work in progress!
 
 ## Installation
 
+Install via CRAN with:
+
 ```{r eval=FALSE}
-devtools::install_github("medewitt/tidyqwi")
+install.packages("tidyqwi")
 ```
+
+Or the development version at:
+
+```{r eval=FALSE}
+remotes::install_github("medewitt/tidyqwi")
+```
+
+## Use
 
 After installation you can load and retrieve the desired data!
 
