### 2019-06-12 #1462 @tweed1e
Hi, sorry for the delay, I've completed my review—I opened a few issues which the author has dealt with, and now I'm satisfied that it passes all requirements. 👍👍.

(The author guidelines say leaving this comment is sufficient, so I'm guessing I shouldn't "Close and comment".)
