### 2019-06-16 #1462 @alexhanna
@medewitt we need the documentation to be located in a `paper.md` file for submission. We need a doc in the format described on the [submission guidelines site](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).
