### 2019-06-16 #1462 @medewitt
Hi @alexhanna thank you for your work in this and thank you @tweed1e for the review.

 I think everything is in compliance to the guidelines at https://github.com/medewitt/tidyqwi/tree/master/inst
