### 2019-06-16 #1462 @alexhanna
I don't think so, although you could try to change that and recheck the PDF. @arfon, is there a particular location in the repository that the `paper.md` must be located?
