### 2019-06-16 #1462 @arfon
I don't think Whedon can find the paper as the repository address won't clone (`https://github.com/medewitt/tidyqwi/tree/master/inst`). I've changed the URL at the top of the thread to be `https://github.com/medewitt/tidyqwi`
