### 2019-06-17 #1462 @alexhanna
@medewitt looks like a few references are missing DOIs. Could you check on those?
