### 2019-06-17 b8988d8
updated references with DOI
