+++ inst/paper.bib
@@ -41,13 +41,18 @@
 }
 
 @article{abowd2012did,
-  title={Did the Housing Price Bubble Clobber Local Labor Market Job and Worker Flows When It Burst?},
-  author={Abowd, John M and Vilhuber, Lars},
-  journal={American Economic Review},
-  volume={102},
-  number={3},
-  pages={589--93},
-  year={2012}
+	title = {Did the {Housing} {Price} {Bubble} {Clobber} {Local} {Labor} {Market} {Job} and {Worker} {Flows} {When} {It} {Burst}?},
+	volume = {102},
+	issn = {0002-8282},
+	url = {https://www.jstor.org/stable/23245599},
+	doi = {10.1257/aer.102.3.589},
+	number = {3},
+	urldate = {2019-05-16},
+	journal = {The American Economic Review},
+	author = {Abowd, John M. and Vilhuber, Lars},
+	year = {2012},
+	pages = {589--593},
+	file = {JSTOR Full Text PDF:C\:\\Users\\dewittme\\Zotero\\storage\\WB68ZGY4\\Abowd and Vilhuber - 2012 - Did the Housing Price Bubble Clobber Local Labor M.pdf:application/pdf}
 }
 
 
@@ -58,25 +63,39 @@
   year={2019}
 }
 
-
 @article{curtis2018loses,
-  title={Who Loses under Cap-and-Trade Programs? The Labor Market Effects of the Nox Budget Trading Program},
-  author={Curtis, E Mark},
-  journal={Review of Economics and Statistics},
-  volume={100},
-  number={1},
-  pages={151--166},
-  year={2018},
-  publisher={MIT Press}
+	title = {Who {Loses} under {Cap}-and-{Trade} {Programs}? {The} {Labor} {Market} {Effects} of the {NO} $_{\textrm{x}}$ {Budget} {Trading} {Program}},
+	volume = {100},
+	issn = {0034-6535, 1530-9142},
+	shorttitle = {Who {Loses} under {Cap}-and-{Trade} {Programs}?},
+	url = {https://www.mitpressjournals.org/doi/abs/10.1162/REST_a_00680},
+	doi = {10.1162/REST_a_00680},
+	abstract = {This paper tests how a major cap-and-trade program, known as the NOx budget trading program (NBP), affected labor markets in the manufacturing sector. The cap-and-trade program dramatically decreased levels of NOx emissions and added substantial costs to regulated ﬁrms. Using a triple-differences approach, I examine how labor markets adjusted in manufacturing industries that were exposed to the program. I ﬁnd that overall employment in the manufacturing sector dropped by 1.3\%, with energy-intensive industries losing up to 4.8\%. Employment declines are shown to have occurred primarily through decreased hiring rates rather than increased separation rates, thus mitigating the impact on incumbent workers. Young workers experienced the largest employment declines, and earnings of newly hired workers fell after the regulation began.},
+	language = {en},
+	number = {1},
+	urldate = {2019-05-16},
+	journal = {The Review of Economics and Statistics},
+	author = {Curtis, E. Mark},
+	month = mar,
+	year = {2018},
+	pages = {151--166},
+	file = {Curtis - 2018 - Who Loses under Cap-and-Trade Programs The Labor .pdf:C\:\\Users\\dewittme\\Zotero\\storage\\P7C7LG6B\\Curtis - 2018 - Who Loses under Cap-and-Trade Programs The Labor .pdf:application/pdf}
 }
 
+
 @article{dube2016minimum,
-  title={Minimum Wage Shocks, Employment flows, and Labor Market Frictions},
[…]
