### 2019-06-17 #1462 @alexhanna
No worries. This looks ready to go then. The last thing we need is for you to create an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the Zenodo/figshare archive, please make sure that:
