### 2019-06-17 #1462 @danielskatz
👋 @arfon - note that there's a problem on line 40 of the [xml](https://github.com/openjournals/joss-papers/pull/769/commits/4876bc02ec4ea60d6be941ded63ef85d057df952) - your thoughts on how to proceed?
