### 2019-06-17 #1462 @arfon
@medewitt - could you change your name entry to `name: Michael E DeWitt Jr` (i.e. remove the comma)?
