### 2019-06-17 #1462 @arfon
This will allow Whedon to parse the name field properly.
