### 2019-06-17 9f33dc0
Update paper.md
remove comma from name
